move authentication process to service class
